requirements.txt
chat_exporter==2.5.0
discord.py==2.1.0
gspread==5.7.1
pendulum==2.1.2
python-dotenv==0.21.1
titlecase==2.4
